+++ src/components/LinesModal/LineAddModal.tsx
@@ -1,11 +1,11 @@
 import React, { ChangeEventHandler, FormEventHandler, useEffect, useMemo, useState } from 'react';
-import { useSelector } from 'react-redux';
 import { LINE, LINE_COLORS, SECTION } from '../../constants/appInfo';
 import { ERROR_MESSAGE } from '../../constants/message';
+import useLines from '../../hooks/useLines';
+import useStations from '../../hooks/useStations';
 import useUpdateEffect from '../../hooks/useUpdateEffect';
 import { addLine } from '../../redux/lineSlice';
-import { loadStations } from '../../redux/stationSlice';
-import { RootState, useAppDispatch } from '../../redux/store';
+import { useAppDispatch } from '../../redux/store';
 import { isKoreanAndNumber } from '../../util/validator';
 import Button from '../@common/Button/Button';
 import ColorRadio from '../@common/ColorRadio/ColorRadio';
@@ -35,8 +35,8 @@ interface ErrorMessage {
 }
 
 const LineAddModal = ({ onClose }: Props): JSX.Element => {
-  const { stations } = useSelector((state: RootState) => state.station);
-  const { lines, errorMessage } = useSelector((state: RootState) => state.line);
+  const { stations } = useStations();
+  const { lines, errorMessage } = useLines();
   const usedLineColors = useMemo(() => lines.map((line) => line.color), [lines]);
   const dispatch = useAppDispatch();
 
@@ -60,12 +60,6 @@ const LineAddModal = ({ onClose }: Props): JSX.Element => {
     section: '',
     distance: '',
   });
-
-  useEffect(() => {
-    if (stations.length === 0) {
-      dispatch(loadStations());
-    }
-  }, []);
 
   useUpdateEffect(() => {
     if (formInput.upStationId === '' || formInput.downStationId === '') {
