+++ src/pages/Stations/Stations.tsx
@@ -1,4 +1,4 @@
-import React, { ChangeEventHandler, FormEventHandler, useEffect, useState } from 'react';
+import React, { ChangeEventHandler, FormEventHandler, useState } from 'react';
 import { useSelector } from 'react-redux';
 import Button from '../../components/@common/Button/Button';
 import CardTemplate from '../../components/@common/CardTemplate/CardTemplate';
@@ -9,9 +9,10 @@ import Loading from '../../components/@common/Loading/Loading';
 import ListItem from '../../components/@shared/ListItem/ListItem';
 import { PAGE_INFO, STATION } from '../../constants/appInfo';
 import { ERROR_MESSAGE } from '../../constants/message';
+import useStations from '../../hooks/useStations';
 import useThemeColor from '../../hooks/useThemeColor';
 import useUpdateEffect from '../../hooks/useUpdateEffect';
-import { addStation, deleteStation, loadStations } from '../../redux/stationSlice';
+import { addStation, deleteStation } from '../../redux/stationSlice';
 import { RootState, useAppDispatch } from '../../redux/store';
 import { isKoreanAndNumber } from '../../util/validator';
 import { StationForm, StationList, StationName, StationNameInput } from './Stations.styles';
@@ -19,18 +20,12 @@ import { StationForm, StationList, StationName, StationNameInput } from './Stati
 const Stations = (): JSX.Element => {
   const themeColor = useThemeColor();
   const isLogin = useSelector((state: RootState) => state.login.isLogin);
-  const { stations, isLoading, errorMessage } = useSelector((state: RootState) => state.station);
+  const { stations, isLoading, errorMessage } = useStations();
   const dispatch = useAppDispatch();
 
   const [formInput, setFormInput] = useState('');
   const [validationErrorMessage, setValidationErrorMessage] = useState('');
   const isValidStationInput = formInput !== '' && validationErrorMessage === '';
-
-  useEffect(() => {
-    if (stations.length === 0) {
-      dispatch(loadStations());
-    }
-  }, []);
 
   useUpdateEffect(() => {
     if (errorMessage === '') {
